Move helpers.py and data from tests/integration to tests/
Unit tests will need to use helpers.py and data/ too, so moving it to
tests/ where helpers.py can be shared.
